refactor: migrate GUI Agent utilities to shared package
- Create multimodal/tarko/shared-utils/src/gui-agent.ts with reusable utilities
- Move convertToGUIResponse, convertToNormalizedAction, createErrorAction to shared package
- Add GUIExecuteResult interface for standardized execute results
- Update browser-gui-agent.ts to use shared utilities
- Enable other Agents to easily adopt GUI Agent type system
- Remove duplicate code and improve maintainability
