+++ multimodal/tarko/shared-utils/src/index.ts
@@ -7,3 +7,4 @@ export * from './logger';
 export * from './deepMerge';
 export * from './env';
 export * from './filter';
+export * from './gui-agent';
